Break BuyClient to small services, and make it to be a facade
