+++ MobileBuy/buy/src/main/java/com/shopify/buy/dataprovider/CheckoutService.java
@@ -38,6 +38,10 @@ import rx.Observable;
  */
 public interface CheckoutService {
 
+    String getWebReturnToUrl();
+
+    String getWebReturnToLabel();
+
     /**
      * Initiate the Shopify checkout process with a new Checkout object.
      *
